SWC-3052: refactor markdown preview into markdown editor widget (so that it's available everywhere).  Also remove bootstrap modal fade animation to speed up app interaction
