+++ src/main/java/org/sagebionetworks/web/client/GlobalApplicationStateView.java
@@ -3,4 +3,5 @@ package org.sagebionetworks.web.client;
 
 public interface GlobalApplicationStateView {
 	void showVersionOutOfDateGlobalMessage();
+	void initGlobalViewProperties();
 }
