+++ src/main/java/org/sagebionetworks/web/client/widget/entity/MarkdownEditorAction.java
@@ -43,6 +43,7 @@ public enum MarkdownEditorAction {
 	INSERT_SYNAPSE_FORM,
 	INSERT_PREVIEW,
 	SET_PROJECT_BACKGROUND,
+	MARKDOWN_PREVIEW,
 	INSERT_TEAM_MEMBERS,
 	BOLD,
 	ITALIC,
