+++ src/main/resources/org/sagebionetworks/web/client/view/LoginViewImpl.ui.xml
@@ -66,7 +66,7 @@
 							<g:Button ui:field="takePledgeButton" styleName="btn btn-lg btn-primary margin-top-10 right" text="Take the pledge, I do"/>
 						</div>
 					</div>
-					<b:Modal ui:field="termsOfUseDialog" title="" closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="false" size="LARGE">
+					<b:Modal ui:field="termsOfUseDialog" title="" closable="true" dataBackdrop="STATIC" dataKeyboard="false" size="LARGE">
 						<b:ModalBody>
 							<g:SimplePanel ui:field="termsOfUseContainer" />
 						</b:ModalBody>
