+++ src/main/resources/org/sagebionetworks/web/client/widget/discussion/modal/DiscussionThreadModalViewImpl.ui.xml
@@ -3,7 +3,7 @@
 	xmlns:g="urn:import:com.google.gwt.user.client.ui"
 	xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
 	xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
-	<b:Modal ui:field="threadModal" closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="false" addStyleNames="modal-fullscreen">
+	<b:Modal ui:field="threadModal" closable="true" dataBackdrop="STATIC" dataKeyboard="false" addStyleNames="modal-fullscreen">
 		<b:ModalBody>
 			<b:TextBox ui:field="threadTitle" placeholder="Title" addStyleNames="font-size-24 margin-bottom-10" height="50px" />
 			<bh:Div ui:field="markdownEditorContainer"></bh:Div>
