+++ src/main/resources/org/sagebionetworks/web/client/widget/discussion/modal/ReplyModalViewImpl.ui.xml
@@ -3,7 +3,7 @@
 	xmlns:g="urn:import:com.google.gwt.user.client.ui"
 	xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
 	xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
-	<b:Modal ui:field="replyModal" closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="false" addStyleNames="modal-fullscreen">
+	<b:Modal ui:field="replyModal" closable="true" dataBackdrop="STATIC" dataKeyboard="false" addStyleNames="modal-fullscreen">
 		<b:ModalBody>
 			<bh:Div ui:field="markdownEditorContainer"></bh:Div>
 			<bh:Div ui:field="synAlertContainer"></bh:Div>
