+++ src/main/resources/org/sagebionetworks/web/client/widget/docker/modal/AddExternalRepoModalViewImpl.ui.xml
@@ -3,7 +3,7 @@
 	xmlns:g="urn:import:com.google.gwt.user.client.ui"
 	xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
 	xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
-	<b:Modal ui:field="addExternalRepoModal" closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="false" addStyleNames="modal">
+	<b:Modal ui:field="addExternalRepoModal" closable="true" dataBackdrop="STATIC" dataKeyboard="false" addStyleNames="modal">
 		<b:ModalBody>
 			<b:Row>
 				<b:Column size="XS_12, MD_9">
