+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/EditProjectMetadataModalViewImpl.ui.xml
@@ -6,7 +6,7 @@
 	xmlns:bg="urn:import:org.gwtbootstrap3.client.ui.gwt" xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
 
 	<!-- Dialog for creating a new table -->
-	<b:Modal ui:field="modal" title="Edit Project Metadata" closable="true" fade="true"
+	<b:Modal ui:field="modal" title="Edit Project Metadata" closable="true"
 		dataBackdrop="STATIC" dataKeyboard="false">
 		<b:ModalBody>
 			<bh:Div>
