+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/EditRegisteredTeamDialogViewImpl.ui.xml
@@ -2,7 +2,7 @@
 <ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
 	xmlns:g="urn:import:com.google.gwt.user.client.ui" xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
 	xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
-	<b:Modal title="Edit Team Registration" closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="false">
+	<b:Modal title="Edit Team Registration" closable="true" dataBackdrop="STATIC" dataKeyboard="false">
 		<b:ModalBody>
 			<b:FormGroup>
 				<b:FormLabel for="recruitment">Team recruitment message</b:FormLabel>
