+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/JavaClientModalWidgetViewImpl.ui.xml
@@ -11,7 +11,7 @@
   
   	<g:HTMLPanel addStyleNames="displayInline">
 	  	<b:ImageAnchor dataToggle="MODAL" dataTarget="#javaClient" url='{sageImageBundle.logoJava45.getSafeUri.asString}'/>
-	  	<b:Modal title="Synapse Java Client" closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="true" b:id="javaClient">
+	  	<b:Modal title="Synapse Java Client" closable="true" dataBackdrop="STATIC" dataKeyboard="true" b:id="javaClient">
 			<b:ModalBody>
 				<e:JavaClientEntityGetterUIWidgetViewImpl ui:field="getEntity"/>
 		  		<e:JavaClientInstallWidgetViewImpl ui:field="install"/>
