+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/RegisterTeamDialogViewImpl.ui.xml
@@ -2,7 +2,7 @@
 <ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
 	xmlns:g="urn:import:com.google.gwt.user.client.ui" xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
 	xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
-	<b:Modal title="Register Team" closable="false" fade="true" dataBackdrop="STATIC"
+	<b:Modal title="Register Team" closable="false" dataBackdrop="STATIC"
 		dataKeyboard="false">
 		<b:ModalBody>
 			<b:FormGroup ui:field="teamSelectionUI">
