+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/annotation/EditAnnotationsDialogViewImpl.ui.xml
@@ -5,7 +5,7 @@
 
 	<g:HTMLPanel>
 		<b:Modal ui:field="editModal" title="Annotations" closable="true"
-			fade="true" dataBackdrop="STATIC" dataKeyboard="false" b:id="editAnnotationsModal">
+			dataBackdrop="STATIC" dataKeyboard="false" b:id="editAnnotationsModal">
 			<b:ModalBody>
 				<t:Table width="80%">
 					<t:TableRow>
