+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/browse/FilesBrowserViewImpl.ui.xml
@@ -22,7 +22,7 @@
 		<g:SimplePanel ui:field="files" styleName="highlight-box padding-top-0-imp" />
 
 		<b:Modal ui:field="newFolderDialog" title="New Folder"
-			closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="false"
+			closable="true" dataBackdrop="STATIC" dataKeyboard="false"
 			size="LARGE">
 			<b:ModalBody>
 				<b:FormGroup>
