+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/controller/ProvenanceEditorWidgetViewImpl.ui.xml
@@ -6,8 +6,7 @@
 	xmlns:bg="urn:import:org.gwtbootstrap3.client.ui.gwt" xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html"
 	xmlns:w="urn:import:org.sagebionetworks.web.client.widget">
 	<g:HTMLPanel>
-	    <b:Modal ui:field="modal" closable="true" fade="true"
-			dataBackdrop="STATIC" dataKeyboard="false" size="LARGE">
+	    <b:Modal ui:field="modal" closable="true" dataBackdrop="STATIC" dataKeyboard="false" size="LARGE">
 			<b:ModalHeader>
 			    <b:Heading size="H4" text="Edit Provenance (Activity)" addStyleNames="displayInline"/>
 			    <w:HelpWidget helpMarkdown="Provenance tracks the relationship between data, code and analytical results."
