+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/controller/SynapseAlertViewImpl.ui.xml
@@ -19,7 +19,7 @@
   		<b:Alert ui:field="alert" type="DANGER" addStyleNames="margin-bottom-0-imp">
   			<bh:Strong ui:field="alertText"></bh:Strong>
   		</b:Alert>
-  		<b:Modal ui:field="jiraDialog" title="Synapse Error" closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="true">
+  		<b:Modal ui:field="jiraDialog" title="Synapse Error" closable="true" dataBackdrop="STATIC" dataKeyboard="true">
 			<b:ModalBody>
 				<b:FieldSet>
 					<b:FormGroup>
