+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/dialog/BaseEditWidgetDescriptorViewImpl.ui.xml
@@ -9,7 +9,7 @@
 	<ui:with field='icons' type='org.sagebionetworks.web.client.IconsImageBundle'/>
  	<ui:with field='sageImageBundle' type='org.sagebionetworks.web.client.SageImageBundle'/>
  	
- 	<b:Modal size="LARGE" closable="true" fade="true" dataBackdrop="STATIC" dataKeyboard="false">
+ 	<b:Modal size="LARGE" closable="true" dataBackdrop="STATIC" dataKeyboard="false">
 	  <b:ModalBody>
 	    <bh:Div styleName="margin-bottom-20">
 	    	<g:SimplePanel ui:field="paramsPanel"></g:SimplePanel>
