+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/editor/APITableColumnManagerViewImpl.ui.xml
@@ -12,7 +12,7 @@
 		<bh:Div ui:field="columnRenderersTable" addStyleNames="form-horizontal"/>
 		<b:Button ui:field="addColumnButton" icon="PLUS" size="SMALL" addStyleNames="margin-top-10">Add Column Renderer</b:Button>
 		<b:Modal ui:field="newColumnModal" title="Column Configuration"
-			fade="true" dataBackdrop="STATIC" dataKeyboard="false">
+			dataBackdrop="STATIC" dataKeyboard="false">
 			<b:ModalBody>
 				<bh:Div addStyleNames="form-horizontal">
 					<b:FieldSet>
