+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/editor/UserSelectorViewImpl.ui.xml
@@ -4,7 +4,7 @@
 	xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
 	xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
 	
-	<b:Modal ui:field="modal" closable="true" fade="true" dataBackdrop="STATIC" title="Find Synapse User">
+	<b:Modal ui:field="modal" closable="true" dataBackdrop="STATIC" title="Find Synapse User">
 		<b:ModalBody>
 			<bh:Div ui:field="suggestBoxContainer"/>
 		</b:ModalBody>
