+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/renderer/WikiSubpagesOrderEditorModalWidgetViewImpl.ui.xml
@@ -6,7 +6,7 @@
 	xmlns:bg="urn:import:org.gwtbootstrap3.client.ui.gwt" xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
 
 	<!-- Dialog for creating a new table -->
-	<b:Modal ui:field="uiModal" title="Subpage Order Editor" closable="true" fade="true"
+	<b:Modal ui:field="uiModal" title="Subpage Order Editor" closable="true"
 		dataBackdrop="STATIC" dataKeyboard="false" size="SMALL">
 		<b:ModalBody>
 			<g:SimplePanel ui:field="editorPanel" />
