+++ src/main/resources/org/sagebionetworks/web/client/widget/login/LoginModalViewImpl.ui.xml
@@ -6,7 +6,7 @@
 	xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
 
 	<!-- Dialog for gathering user credentials -->
-	<b:Modal closable="true" fade="true"
+	<b:Modal closable="true"
 		dataBackdrop="STATIC" dataKeyboard="false">
 		<b:ModalBody>
 		    <bh:Text ui:field="instructions"></bh:Text>
