+++ src/main/resources/org/sagebionetworks/web/client/widget/sharing/AccessControlListModalWidgetViewImpl.ui.xml
@@ -7,7 +7,7 @@
 	xmlns:w="urn:import:org.sagebionetworks.web.client.widget">
 
 	<!-- Dialog for creating a new table -->
-	<b:Modal ui:field="uiModal" fade="true"
+	<b:Modal ui:field="uiModal"
 		dataBackdrop="STATIC" dataKeyboard="false" size="LARGE">
 		<b:ModalHeader>
 			<b:Heading size="H4" text="Sharing Settings" addStyleNames="displayInline" />
