+++ src/main/resources/org/sagebionetworks/web/client/widget/table/modal/fileview/ScopeWidgetViewImpl.ui.xml
@@ -10,7 +10,7 @@
 		<b:Button ui:field="editButton" icon="EDIT" addStyleNames="margin-top-10" visible="false" text="Edit Scope" type="INFO"/>
 		
 		<b:Modal ui:field="editModal" title="Edit Scope" closable="true"
-			fade="true" dataBackdrop="STATIC" dataKeyboard="false" b:id="editScopeModal"
+			dataBackdrop="STATIC" dataKeyboard="false" b:id="editScopeModal"
 			size="LARGE">
 			<b:ModalBody>
 				<g:SimplePanel ui:field="editScopeContainer"/>
