+++ src/main/resources/org/sagebionetworks/web/client/widget/table/modal/wizard/ModalWizardViewImpl.ui.xml
@@ -6,7 +6,7 @@
 	xmlns:bg="urn:import:org.gwtbootstrap3.client.ui.gwt" xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
 
 	<!-- Dialog for creating a new table -->
-	<b:Modal closable="true" fade="true"
+	<b:Modal closable="true"
 		dataBackdrop="STATIC" dataKeyboard="false">
 		<b:ModalBody>
 			<bh:Text ui:field="instructions"></bh:Text>
