+++ src/main/resources/org/sagebionetworks/web/client/widget/table/v2/schema/ColumnModelsViewBaseImpl.ui.xml
@@ -7,7 +7,7 @@
 		<g:SimplePanel ui:field="viewerPanel" />
 		<!-- This modal dialog will contain the editor view -->
 		<b:Modal ui:field="editModal" title="Edit Columns" closable="true"
-			fade="true" dataBackdrop="STATIC" dataKeyboard="false" b:id="editColumnsModal"
+			dataBackdrop="STATIC" dataKeyboard="false" b:id="editColumnsModal"
 			size="LARGE">
 			<b:ModalBody>
 				<g:SimplePanel ui:field="columnEditorModalPanel"/>
