+++ src/main/resources/org/sagebionetworks/web/client/widget/team/controller/TeamDeleteModalWidgetViewImpl.ui.xml
@@ -5,7 +5,7 @@
 	xmlns:a="urn:import:org.sagebionetworks.web.client.widget.table.v2"
 	xmlns:bg="urn:import:org.gwtbootstrap3.client.ui.gwt" xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
 
-	<b:Modal ui:field="modal" title="Delete Team?" closable="true" fade="true"
+	<b:Modal ui:field="modal" title="Delete Team?" closable="true"
 		dataBackdrop="STATIC" dataKeyboard="false">
 		<b:ModalBody>
 		    <bh:Text>Are you sure you want to delete this Team?</bh:Text>
