+++ src/main/resources/org/sagebionetworks/web/client/widget/team/controller/TeamLeaveModalWidgetViewImpl.ui.xml
@@ -5,7 +5,7 @@
 	xmlns:a="urn:import:org.sagebionetworks.web.client.widget.table.v2"
 	xmlns:bg="urn:import:org.gwtbootstrap3.client.ui.gwt" xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html">
 
-	<b:Modal ui:field="modal" title="Leave Team" closable="true" fade="true"
+	<b:Modal ui:field="modal" title="Leave Team" closable="true"
 		dataBackdrop="STATIC" dataKeyboard="false">
 		<b:ModalBody>
 		    <bh:Text>You will lose access to resources shared with the team. Are you sure?</bh:Text>
