+++ src/main/webapp/Portal.css
@@ -722,20 +722,20 @@ a.shortBtn {
 .margin-left-1 {
 	margin-left: 1px;
 }
-.margin-left-2 {
+.margin-left-2, #rootPanel .margin-left-2 {
 	margin-left: 2px;
 }
 .margin-left-3 {
 	margin-left: 3px;
 }
 
-.margin-left-5 {
+.margin-left-5, #rootPanel .margin-left-5 {
 	margin-left: 5px;
 }
 .margin-left-9 {
 	margin-left: 9px;
 }
-.margin-left-10 {
+.margin-left-10, #rootPanel .margin-left-10{
 	margin-left: 10px;
 }
 .margin-left-12 {
@@ -2471,6 +2471,10 @@ td.datePickerMonth {
 
 .border-bottom-1 {
   border-bottom: 1px solid #dddddd;
+}
+
+.border-left-1 {
+  border-left: 1px solid #dddddd;
 }
 
 /* Provenance graph css */
